+++ README.md
@@ -111,9 +111,9 @@ pixiv及び本ソフトウェアに登場する用語について説明します
 #### ルールの記法詳細
 * コメント行の規則
 	* 空行は無視される。
-	* # で始まる行はコメント行とみなされ、無視される。
-	* 整形のために # の前に空白文字を入れてもよい。前に来るのが空白文字でなければエラーとなる。
-	* ルール行で # を用いることはできない。
+	* `#` で始まる行はコメント行とみなされ、無視される。
+	* 整形のために `#` の前に空白文字を入れてもよい。前に来るのが空白文字でなければエラーとなる。
+	* ルール行でコメントのために `#` を用いることはできない。
 * ルール文の規則
 	* ルール文の区切りは空白文字である。
 	* 整形のために区切り文字として空白文字を連続して使用してもよい。
