+++ test/rule_config_parser.test.mjs
@@ -1,0 +1,34 @@
+import {
+  RuleConfigParser
+} from '../src/domain.mjs';
+
+import assert from 'assert';
+
+describe('RuleConfigParser', () => {
+
+  const parser = new RuleConfigParser();
+
+  describe('#parse', () => {
+
+    describe('when collect rule is given', () => {
+
+      it('should parse rule', () => {
+        const rule =
+          '# 非公開設定\n' +
+          'private R-18 R-18G R-17.9 R-15\n' +
+          '# 一般\n' +
+          'pattern オリジナル オリジナル\n' +
+          '# 艦これ\n' +
+          'pattern_all ~1 ^艦これ$|^艦隊これくしょん$ ^(.+)(改|改二)$\n' +
+          'match 卯月 うーちゃん\n' +
+          '# 東方\n' +
+          'match_all 多々良小傘 東方 小傘\n';
+          // 'addition_pattern_all ~1(アズールレーン) ^アズールレーン|アズレン|碧蓝航线$ ^睦月|如月|卯月|水無月)$';
+
+        console.log(parser.parse(rule));
+        assert(true);
+
+      });
+    });
+  });
+});
